Native style/Button: Avoid that the label gets elided
This was observed on Button on Windows style, where the labels size
hint was a fractional number (e.g. 38.26). This got truncated by
QQuickStyleItem::contentSize(), which caused the content item to get
less space than it actually needed

It was only observed on Button, but it was likely to also happen
with other control types.

Change-Id: I50b98fa7d54e5be2cedeeaf60018367689f168cf
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>