Remove some outdated comments about FontLoader.name
In 15ea475b40f6ad28d46e5cbd65a1ccc8556a53df, I made FontLoader.name
readonly and got rid of the usage where it could also be used
as an indirection for the font family string. But there were a
couple of comments about this usage still in the documentation,
so those should also be removed.

Change-Id: Ibf182e1252644e155843dea8f85f4a2f7fb56297
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>