Fix tst_QQuickText::fontSizeMode
The test hardcodes the height of the Text element to be 35,
and then it does a pre-check that this is higher than the
height of the font. This is obviously a bit flaky and caused
an error because the font at the given size was 35.7 and was
rounded up.

I did attempt to make the logic a bit more dynamic, and have the
object's initial height be based on the height of the font, but
other parts of the test would then fail because they depend on
the size of the object.

In the end, the simplest way seemed to be to just decrease the
font size until it fit inside the original Text height and
leave it at that.

Fixes: QTBUG-88207
Change-Id: I9c327806bde8c339b299302004dfb72b34e87bcd
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>