Rename "Porting" pages to "Changes"
The content in the porting guides are closer to a changelog than
a porting guide. At this point, it is easier for maintainers and
contributors to write in a changelog than a guide. This change
should help with readability and is closer to the usage of
"Changes" in documentation.

Part of a rename in other submodules.

Task-number: QTBUG-87156
Change-Id: Ib34fc9e3e5719c75fcd7d18d9d2cc7dd994c3116
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>