Rename "Porting" pages to "Changes"
The content in the porting guides are closer to a changelog than
a porting guide. At this point, it is easier for maintainers and
contributors to write in a changelog than a guide. This change
should help with readability and is closer to the usage of
"Changes" in documentation.

Part of a rename in other submodules.

Task-number: QTBUG-87156
Change-Id: Ibb58a3cba5394962a5709aeb08e3701e16c7775d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>