Add simpler non point particle level
Add one more PerformanceLevel which is similar to Color, except it
draws using triangles instead of points. This is optimal for d3d11
which doesn't support VertexShaderPointSize.

Also rename first levels to *Point to be clear.

Testing this on my windows laptop shows ~15% decrease in RAM usage
for imageparticle/colored example.

Task-number: QTBUG-88294
Change-Id: Ib93bc9fd0d2ef4f1cb8e0d9dcba0db652f288398
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>