Follow up on change of QQmlListProperty to qsizetype
This deals with d7008c79d4ec023527ebfc118ad47f40075f244d in
qtdeclarative which aligned QQmlListProperty with our container types.
Note: not changing the respective APIs of the QQuick classes
in this change. Ie. QQuickPlatformMenu::insertItem still operates on
int as index, and QQuickMenuPrivate still stores currentIndex as int.
This might need to be addressed in a follow-up commit.

Fixes: QTBUG-88362
Change-Id: Ia663cfa47fa91c55997cdef288b2a866107a5f25
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>