QQuickWindow: better hover debug, and a reminder
In the qt.quick.hover.trace category, the position is the most
important thing for now.  The output for "q" is verbose and usually
there's only one window anyway, so just put the title last, in case
we need to debug a multi-window scenario.

Dealing with hover in multi-device scenarios is going to be interesting
one of these days.

Change-Id: I2b687085432ce2e02ca764b8b4669282e0180c54
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>