Add QtQuick.Layout dependency to native style qmldir
Without it qmlimportscanner misses the dependency of FocusFrame.qml
and static builds fail to add the dependency leading to a QComponent is
not ready error.
This happens because the qml file is in a qrc, and thus not visible to qmlimportscanner.
Adding a depends QtQuick.Layouts fixes it  (and updating the controls
dependency to drop the version)

Fixes: QTBUG-88492
Change-Id: I9be3e26c3a03b663cb90618368f3480f141a6b04
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>