macOS: don't use a focus frame around TextArea
As it turns out, text areas in general should not show a
focus frame. We don't do that for widgets, and AppKit
doesn't do that for NSTextView. And for good reason, we
don't want to show a big focus ring around the whole text
edit in e.g a text editor application.

Change-Id: Idc344c9f8d9f4a11dce22b2d2284800f05cd551a
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>