Re-enable the caching of shader data loaded from files
There was a mismatch for the cache key: sometimes using the original url
as specified in the QML code, sometimes using the resolved url (which has,
for example, a relative path expanded to absolute). Change this to be
consistent.

Fixes: QTBUG-88673
Change-Id: I201750716d3ba6dbe73a4799ac56f26f9b8ec820
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit f95b99902fae6e53a9da5fbccb6e1616299fb3bf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>