Fix multipointtoucharea_interop tests
In several places these were doing stuff like

    touch.press(1, p1).commit();
    QTRY_VERIFY(devPriv->pointById(0)->passiveGrabbers.contains(drag));

The point ID is really the key in activePoints, it doesn't necessarily
start from 0; so pointById() has to be given the same point ID that was
used in the event, otherwise it will construct a new point(0).

But now tst_MptaInterop::touchesThenPinch fails further down: after the
complicated series of events, when only one point is still being dragged,
DragHandler fails to grab it and start dragging for some reason.

dragHandlerInParentStealingGrabFromItem() uses mouse not touch.

Pick-to: 6.0
Task-number: QTBUG-86729
Change-Id: I7ad0d3f30cb5fee15fe83183d62ee10a020c14df
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>