QQuickParticleAffector: Fix warnings about copying non-shared containers
Use range-based for, fixing:
'QtPrivate::warnIfContainerIsNotShared': Do not use foreach/Q_FOREACH with containers which are not implicitly shared. Prefer using a range-based for loop with these containers: `for (const auto &it : container)`, keeping in mind that range-based for doesn't copy the container as Q_FOREACH does
            T=QVarLengthArray<QQuickParticleGroupData *,32> &
'QtPrivate::warnIfContainerIsNotShared': Do not use foreach/Q_FOREACH with containers which are not implicitly shared. Prefer using a range-based for loop with these containers: `for (const auto &it : container)`, keeping in mind that range-based for doesn't copy the container as Q_FOREACH does
            T=QQuickParticleVarLengthArray<QQuickParticlePainter *,4> &

Change-Id: I88d7405abd89d2fa47549623997cbc02a9bd1145
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 938986fb464601ad3104ee86d03f2ef562e24315)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>