Windows, CT_LineEdit: don't take dpr into account when calculating geometry
The geometry calculated by the style should not take dpr
into account. The nativestyle framework (and the scenegraph)
understands that all geometry returned by QStyle will need
to be scaled according to the dpr. We will for example
set a dpr on the QPainter (or the QImage it draws on) before
we ask QStyle to draw. This means that the style should not
scale the geometry used in the drawing calls to QPainter since
QPainter will do this automatically.

Task-number: QTBUG-89006
Pick-to: 6.0 6.0.0
Change-Id: Idfcc6690e0ae0062ca00cb829510d2b68c26ebd4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>