Find qml imports inside subdirectories of symlinks
Added QDirIterator::FollowSymlinks flag in function findQmlImportsInDirectory. This is to find qml imports in subdirectories of symlinks. Reason is that qwindeployqt does not scan imports inside directories following symlinks and therefore might skip dependencies that are needed for the application. And this is caused by the qmlimportscanner.

[ChangeLog][tools][qmlimportscanner]
Follow sybmolic links in findQmlImportsInDirectory

Change-Id: I9739164e50cf48ec9e67f63a110e99da83facc2c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>