Quick: Fortify QQuickApplication ctor
If the qApp we have is in fact not a QGuiApplication we cannot
connect those signals. The "if (qApp)" was probably supposed to check
for this but actually doesn't because qApp is a static_cast.

Change-Id: Iaa30e3d56768e97601d51d30500fc3b12d0626a3
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 8a555cc22afb93dbe0e0da83f211c6e4f032bd09)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>