QQmlValueTypeFactory: remove bespoke Qt type cache
QQmlValueTypeFactory contained an array of size QMetaType::User, to
avoid locking for accessing user types.
In Qt 6, User is however significantly larger, causing the array to take
roughly 500 KByte of memory. We therefore unify the handling of user
types and Qt built-in types, by storing all of them in a QHash. We
replace the mutex with a read-write lock to still allow concurrent
read access to the mapping (now for all types and not only Qt builtin
types).

Change-Id: Ie22786b0f33cf27157b151c75d5bb21579a566f6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>